+++ internal/services/azapi_resource_action_resource_test.go
@@ -22,6 +22,18 @@ func TestAccActionResource_basic(t *testing.T) {
 	})
 }
 
+func TestAccActionResource_registerResourceProvider(t *testing.T) {
+	data := acceptance.BuildTestData(t, "azapi_resource_action", "test")
+	r := ActionResource{}
+
+	data.DataSourceTest(t, []resource.TestStep{
+		{
+			Config: r.providerAction(),
+			Check:  resource.ComposeTestCheckFunc(),
+		},
+	})
+}
+
 func TestAccActionResource_providerAction(t *testing.T) {
 	data := acceptance.BuildTestData(t, "azapi_resource_action", "test")
 	r := ActionResource{}
@@ -59,7 +71,7 @@ resource "azapi_resource_action" "test" {
 `, GenericResource{}.defaultTag(data))
 }
 
-func (r ActionResource) providerAction() string {
+func (r ActionResource) registerResourceProvider() string {
 	return `
 provider "azurerm" {
   features {}
@@ -72,6 +84,26 @@ resource "azapi_resource_action" "test" {
   resource_id = "/subscriptions/${data.azurerm_client_config.current.subscription_id}/providers/Microsoft.Compute"
   action      = "register"
   method      = "POST"
+}
+`
+}
+
+func (r ActionResource) providerAction() string {
+	return `
+provider "azurerm" {
+  features {}
+}
+
+data "azurerm_client_config" "current" {}
+
+resource "azapi_resource_action" "test" {
+  type        = "Microsoft.Cache@2023-04-01"
+  resource_id = "/subscriptions/${data.azurerm_client_config.current.subscription_id}/providers/Microsoft.Cache"
+  action      = "CheckNameAvailability"
+  body = jsonencode({
+    type = "Microsoft.Cache/Redis"
+    name = "cacheName"
+  })
 }
 `
 }