+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ENHANCEMENTS:
 - `azapi_resource_action`: Support provider action.
 
 BUG FIXES:
+- Fix a bug that resource id for type `Microsoft.Resources/providers` is not parsed correctly.
 - Fix a bug that resource id for type `Microsoft.Resources/tenants` is not parsed correctly.
 
 ## v1.7.0